Move logdriver to internal under daemon/logger
The logdriver package should be moved to another external repository for
building plugins. While transitioning, move to internal package using
the same structure.
