+++ daemon/logger/internal/logdriver/entry.proto
