+++ daemon/logger/internal/logdriver/gen.go
