+++ daemon/logger/internal/logdriver/io.go
