+++ daemon/logger/local/local.go
@@ -9,11 +9,11 @@ import (
 	"time"
 
 	"github.com/docker/docker/daemon/logger"
+	"github.com/docker/docker/daemon/logger/internal/logdriver"
 	"github.com/docker/docker/daemon/logger/loggerutils"
 	"github.com/docker/docker/daemon/server/backend"
 	"github.com/docker/docker/errdefs"
 	"github.com/docker/go-units"
-	"github.com/moby/moby/api/types/plugins/logdriver"
 	"github.com/pkg/errors"
 )
