+++ daemon/logger/local/read.go
@@ -8,9 +8,9 @@ import (
 	"io"
 
 	"github.com/docker/docker/daemon/logger"
+	"github.com/docker/docker/daemon/logger/internal/logdriver"
 	"github.com/docker/docker/daemon/logger/loggerutils"
 	"github.com/docker/docker/errdefs"
-	"github.com/moby/moby/api/types/plugins/logdriver"
 	"github.com/pkg/errors"
 )
