+++ daemon/logger/plugin.go
@@ -7,10 +7,10 @@ import (
 	"path/filepath"
 
 	"github.com/docker/docker/daemon/internal/stringid"
+	"github.com/docker/docker/daemon/logger/internal/logdriver"
 	"github.com/docker/docker/errdefs"
 	"github.com/docker/docker/pkg/plugingetter"
 	"github.com/docker/docker/pkg/plugins"
-	"github.com/moby/moby/api/types/plugins/logdriver"
 	"github.com/pkg/errors"
 )
